requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==3.4.7
idna==3.2
loguru==0.5.3
lxml==4.6.3
msedge-selenium-tools==3.141.3
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
pynput==1.7.3
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
win32-setctime==1.0.3
